good update which allow each docker instance to process a single las entry
