+++ lidar-processing/.gitignore
@@ -1,0 +1,2 @@
+#this is docker temporary folder
+dockermount_tmp/
