+++ lidar-processing/batch-processing/requirement.txt
@@ -5,3 +5,4 @@ numpy
 pandas
 shapely
 requests
+argparse
